+++ scouting-openid-connect.php
@@ -7,16 +7,16 @@ if ( ! defined( 'ABSPATH' ) ) exit; // Exit if accessed directly
  * @category   Scouting OpenID Connect
  * @package    scouting-openid-connect
  * @author     Job van Koeveringe <job.van.koeveringe@scouting.nl>
- * @copyright  2025 Scouting Nederland
+ * @copyright  2026 Scouting Nederland
  * @license    GPLv3
- * @version    2.1.0
+ * @version    2.2.0
  * @link       https://github.com/Scouting-nl/scouting-openid-connect
  *
  * @wordpress-plugin
  * Plugin Name:          Scouting OpenID Connect
  * Plugin URI:           https://github.com/Scouting-nl/scouting-openid-connect
  * Description:          WordPress plugin for logging in with Scouting Nederland OpenID Connect Server.
- * Version:              2.1.0
+ * Version:              2.2.0
  * Requires at least:    6.6.0
  * Requires PHP:         8.2
  * Author:               Job van Koeveringe